pyproject.toml
[tool.poetry]
name = "email-remote-access"
version = "0.0.0"
description = "Control remote computers using email"
authors = ["Duong Vinh <tienvinh.duong4@gmail.com>"]
package-mode = false

[tool.poe.tasks]
dev = "python cli.py"
lint = "ruff check --fix"
pretty = "ruff format"
demo = "python assets/demo.py"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true

[tool.ruff]
extend-include = ["*.ipynb"]
extend-exclude = ["pkg"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.poetry.dependencies]
python = "^3.12"
environs = "^9.5.0"
pillow = "^10.3.0"
psutil = "^5.9.4"
typer = "^0.7.0"
python-socketio = {extras = ["client"], version = "^5.8.0"}
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
seedir = {extras = ["emoji"], version = "^0.4.2"}
dearpygui = "^1.11.0"
google-api-python-client = "^2.85.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
tabulate = "^0.9.0"
pywinctl = "^0.4"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.26.1"
mypy = "^1.10.1"
pre-commit = "^3.7.1"
ruff = "^0.4.10"
types-tabulate = "^0.9.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
